basic changes needed for central release
