+++ .gitignore
@@ -15,3 +15,5 @@ target
 .swp
 .*.swp
 
+/.settings/
+/LICENSE.txt
