+++ markdown-plugin/pom.xml
@@ -5,16 +5,30 @@
 		<artifactId>nanohttpd-project</artifactId>
 		<version>2.1.0-SNAPSHOT</version>
 	</parent>
-	<artifactId>nanohttpd-webserver</artifactId>
+	<artifactId>nanohttpd-webserver-markdown-plugin</artifactId>
 	<packaging>jar</packaging>
-	<name>NanoHttpd-Webserver</name>
+	<name>NanoHttpd-Webserver-Markdown-Plugin</name>
+	<url>https://github.com/NanoHttpd/nanohttpd</url>
 	<dependencies>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
 			<artifactId>nanohttpd</artifactId>
 			<version>${project.version}</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>nanohttpd-webserver</artifactId>
+			<version>${project.version}</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.pegdown</groupId>
+			<artifactId>pegdown</artifactId>
+			<version>1.4.1</version>
 		</dependency>
 	</dependencies>
+
 	<build>
 		<plugins>
 			<plugin>
