+++ markdown-plugin/src/main/resources/META-INF/services/fi.iki.elonen.WebServerPluginInfo
