+++ webserver/.gitignore
@@ -1,0 +1,2 @@
+/.settings/
+/LICENSE.txt
